requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
requests==2.32.0
python-dotenv==1.0.1